refactor(wizard): consolidate components pages into single file\n\n- New WizardComponentsPages.swift combines Karabiner + Kanata components pages\n- Delete old individual files; no logic changes intended\n- Build succeeds; core tests runner shows no new failures
